httpd: use size_t for buffer size results
The value can never be negative and it simplifies code in places.

We also need to change the i_body which is used as a pointer in some cases.

The value can never be negative and it simplifies code in places.

We also need to change the i_body which is used as a pointer in some cases.

+++ include/vlc_httpd.h
@@ -99,7 +99,7 @@ typedef struct httpd_message_t
 
     /* body */
     int64_t i_body_offset;
-    int     i_body;
+    size_t  i_body;
     uint8_t *p_body;
 
 } httpd_message_t;
@@ -121,13 +121,13 @@ VLC_API char* httpd_ServerIP( const httpd_client_t *cl, char *, int * );
 
 typedef struct httpd_file_t     httpd_file_t;
 typedef struct httpd_file_sys_t httpd_file_sys_t;
-typedef int (*httpd_file_callback_t)( httpd_file_sys_t *, httpd_file_t *, uint8_t *psz_request, uint8_t **pp_data, int *pi_data );
+typedef int (*httpd_file_callback_t)( httpd_file_sys_t *, httpd_file_t *, uint8_t *psz_request, uint8_t **pp_data, size_t *pi_data );
 VLC_API httpd_file_t * httpd_FileNew( httpd_host_t *, const char *psz_url, const char *psz_mime, const char *psz_user, const char *psz_password, httpd_file_callback_t pf_fill, httpd_file_sys_t * ) VLC_USED;
 VLC_API httpd_file_sys_t * httpd_FileDelete( httpd_file_t * );
 
 
 typedef struct httpd_handler_t  httpd_handler_t;
-typedef int (*httpd_handler_callback_t)( void *, httpd_handler_t *, char *psz_url, uint8_t *psz_request, int i_type, uint8_t *p_in, int i_in, char *psz_remote_addr, char *psz_remote_host, uint8_t **pp_data, int *pi_data );
+typedef int (*httpd_handler_callback_t)( void *, httpd_handler_t *, char *psz_url, uint8_t *psz_request, int i_type, uint8_t *p_in, int i_in, char *psz_remote_addr, char *psz_remote_host, uint8_t **pp_data, size_t *pi_data );
 VLC_API httpd_handler_t * httpd_HandlerNew( httpd_host_t *, const char *psz_url, const char *psz_user, const char *psz_password, httpd_handler_callback_t pf_fill, void * ) VLC_USED;
 VLC_API void * httpd_HandlerDelete( httpd_handler_t * );
